requirements.txt
biopython==1.76
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lxml==4.4.2
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
PyQt5==5.13.1
PyQt5-sip==12.7.0
PyQt5-stubs==5.13.1.4
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.7
xlrd==1.2.0
XlsxWriter==1.2.7
xmltodict==0.12.0
